+++ edge-modules/edge-dashboard/Cargo.toml
@@ -16,7 +16,7 @@ os_info = "1.1.1"
 reqwest = "0.9.18"
 serde = "1.0"
 serde_derive = "1.0"
-serde_json = "1.0"
+serde_json = "1.0.40"
 openssl = { version="0.10" }
 structopt = "0.2.16"
 tokio = "0.1.22"
