+++ edge-modules/edge-dashboard/src/lib.rs
@@ -2,32 +2,23 @@
 
 mod error;
 mod health;
+mod modules;
 mod settings;
 mod state;
+mod status;
 
 #[cfg(windows)]
 use std::env;
-use std::net::TcpStream;
 use std::path::{Path, PathBuf};
-use std::process::Command;
 use std::sync::Arc;
 
 use actix_cors::Cors;
-use actix_web::error::ErrorInternalServerError;
-use actix_web::Error as ActixError;
 use actix_web::*;
-use bytes::BytesMut;
+use edgelet_core::Provisioning;
 use edgelet_core::RuntimeSettings;
-use edgelet_core::{LogOptions, Module as EdgeModule, ModuleRuntime, Provisioning};
 use edgelet_docker::Settings as DockerSettings;
-use edgelet_http_mgmt::*;
-use futures::future::{ok, Either, IntoFuture};
-use futures::stream::Stream;
-use futures::Async;
-use futures::Future;
 use serde_derive::Deserialize;
 use structopt::StructOpt;
-use url::Url;
 
 pub use error::Error;
 use settings::Settings;
@@ -46,26 +37,6 @@ impl Context {
             edge_config,
             settings,
         }
-    }
-}
-
-#[derive(Debug)]
-pub struct Module {
-    name: String,
-    status: String,
-}
-
-impl Module {
-    pub fn new(name: String, status: String) -> Self {
-        Module { name, status }
-    }
-
-    pub fn name(&self) -> &String {
-        &self.name
-    }
-
-    pub fn status(&self) -> &String {
-        &self.status
     }
 }
 
@@ -96,13 +67,15 @@ impl Main {
                 .wrap(Cors::new().send_wildcard())
                 .register_data(context.clone())
                 .register_data(device.clone())
-                .service(web::resource("/api/modules/{id}/restart").to_async(restart_module))
-                .service(web::resource("/api/modules/{id}/logs").to_async(get_logs))
-                .service(web::resource("/api/modules").to_async(get_modules))
-                .service(web::resource("/api/provisioning-state").to(get_state))
-                .service(web::resource("/api/connectivity").to(get_connectivity))
-                .service(web::resource("/api/health").to_async(get_health))
-                .service(web::resource("/api/diagnostics").to(get_diagnostics))
+                .service(
+                    web::resource("/api/modules/{id}/restart").to_async(modules::restart_module),
+                )
+                .service(web::resource("/api/modules/{id}/logs").to_async(modules::get_logs))
+                .service(web::resource("/api/modules").to_async(modules::get_modules))
+                .service(web::resource("/api/health").to_async(modules::get_health))
+                .service(web::resource("/api/provisioning-state").to(status::get_state))
+                .service(web::resource("/api/connectivity").to(status::get_connectivity))
+                .service(web::resource("/api/diagnostics").to(status::get_diagnostics))
         })
         .bind(address)?
         .run()?;
@@ -153,264 +126,6 @@ fn set_up(context: web::Data<Arc<Context>>) -> Option<state::Device> {
             String::new(),
         ))
     }
-}
-
-fn get_state(device: web::Data<Option<state::Device>>) -> HttpResponse {
-    if let Some(dev) = device.get_ref() {
-        state::return_response(&dev)
-    } else {
-        HttpResponse::UnprocessableEntity().body("Device connection string unable to be processed.")
-    }
-}
-
-fn get_modules(
-    context: web::Data<Arc<Context>>,
-    info: web::Query<AuthRequest>,
-) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>> {
-    let api_ver = &info.api_version;
-    return_modules(context, api_ver, module_response)
-}
-
-fn get_connectivity(_req: HttpRequest, device: web::Data<Option<state::Device>>) -> HttpResponse {
-    if let Some(dev) = device.get_ref() {
-        if let Some(iothub) = dev.hub_name() {
-            let iothub_hostname = &format!("{}.azure-devices.net", iothub);
-
-            let r = resolve_and_tls_handshake(&(&**iothub_hostname, 443), iothub_hostname);
-            match r {
-                Ok(_) => HttpResponse::Ok().body(""),
-                Err(_) => HttpResponse::UnprocessableEntity()
-                    .body("Failed to establish connection with IoT Hub."),
-            }
-        } else {
-            HttpResponse::UnprocessableEntity().body("IoT Hub name could not be processed")
-        }
-    } else {
-        HttpResponse::UnprocessableEntity().body("IoT Hub name could not be processed")
-    }
-}
-
-fn get_health(
-    context: web::Data<Arc<Context>>,
-    info: web::Query<AuthRequest>,
-) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>> {
-    let api_ver = &info.api_version;
-    return_modules(context, api_ver, health_response)
-}
-
-fn restart_module(
-    req: HttpRequest,
-    context: web::Data<Arc<Context>>,
-    info: web::Query<AuthRequest>,
-) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>> {
-    let module_id = req.match_info().get("id");
-    let api_ver = &info.api_version;
-    let response = context
-        .edge_config
-        .as_ref()
-        .map(|config| {
-            let mgmt_uri = config.connect().management_uri();
-            Either::A(
-                Url::parse(&format!("{}/modules/?api-version={}", mgmt_uri, api_ver))
-                    .map_err(ErrorInternalServerError)
-                    .and_then(|url| ModuleClient::new(&url).map_err(ErrorInternalServerError))
-                    .map(|mod_client| {
-                        if let Some(id) = module_id {
-                            mod_client.restart(id)
-                        } else {
-                            mod_client.restart("")
-                        }
-                        .map(|_| HttpResponse::Ok().body(""))
-                        .map_err(ErrorInternalServerError)
-                    })
-                    .into_future()
-                    .flatten(),
-            )
-        })
-        .unwrap_or_else(|err| {
-            Either::B(ok(HttpResponse::ServiceUnavailable()
-                .content_type("text/plain")
-                .body(format!("{:?}", err))))
-        });
-
-    Box::new(response)
-}
-
-fn get_logs(
-    req: HttpRequest,
-    context: web::Data<Arc<Context>>,
-    info: web::Query<AuthRequest>,
-) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>> {
-    let module_id = req.match_info().get("id").unwrap_or("");
-    let api_ver = &info.api_version;
-
-    let response = context
-        .edge_config
-        .as_ref()
-        .map(move |config| {
-            let mgmt_uri = config.connect().management_uri();
-            Either::A(
-                Url::parse(&format!("{}/modules/?api-version={}", mgmt_uri, api_ver))
-                    .map_err(ErrorInternalServerError)
-                    .and_then(|url| ModuleClient::new(&url).map_err(ErrorInternalServerError)) // can't connect to the endpoint
-                    .map(move |mod_client| {
-                        mod_client
-                            .logs(module_id, &LogOptions::new())
-                            .map_err(ErrorInternalServerError)
-                            .and_then(|data| {
-                                data.map_err(ErrorInternalServerError)
-                                    .fold(BytesMut::new(), |mut acc, chunk| {
-                                        acc.extend_from_slice(chunk.as_ref());
-                                        Ok::<_, ActixError>(acc)
-                                    })
-                                    .and_then(|body| {
-                                        let mut clone = body.freeze().to_vec().clone();
-                                        let mut curr = true;
-                                        let mut next = true;
-                                        // if the next byte after the \0 encodes the length of msg, ignore
-                                        clone.retain(|&byte| {
-                                            curr = next;
-                                            next = byte as char != '\0';
-                                            byte != 0 && byte != 1 && curr && byte as char != '+'
-                                        });
-                                        clone.retain(|&byte| (byte as char).is_ascii());
-                                        if let Ok(content) = String::from_utf8(clone) {
-                                            HttpResponse::Ok().body(content)
-                                        } else {
-                                            HttpResponse::ServiceUnavailable()
-                                                .body("Logs unable to be displayed")
-                                        }
-                                    })
-                            })
-                    })
-                    .into_future()
-                    .flatten(),
-            )
-        })
-        .unwrap_or_else(|err| {
-            Either::B(ok(HttpResponse::ServiceUnavailable()
-                .content_type("text/plain")
-                .body(format!("{:?}", err))))
-        });
-
-    Box::new(response)
-}
-
-fn get_diagnostics() -> HttpResponse {
-    let output = Command::new("iotedge")
-        .args(&["check", "--output", "json"])
-        .output()
-        .expect("Failed to execute command");
-
-    HttpResponse::Ok().body(output.stdout)
-}
-
-fn return_modules(
-    context: web::Data<Arc<Context>>,
-    api_ver: &str,
-    f: fn(Vec<Module>) -> HttpResponse,
-) -> Box<dyn Future<Item = HttpResponse, Error = ActixError>> {
-    let response = context
-        .edge_config
-        .as_ref()
-        .map(move |config| {
-            let mgmt_uri = config.connect().management_uri();
-            Either::A(
-                Url::parse(&format!("{}/modules/?api-version={}", mgmt_uri, api_ver))
-                    .map_err(ErrorInternalServerError)
-                    .and_then(|url| ModuleClient::new(&url).map_err(ErrorInternalServerError))
-                    .map(|mod_client| {
-                        mod_client
-                            .list()
-                            .map(move |data| {
-                                let mods: Vec<Module> = data
-                                    .iter()
-                                    .map(move |c| {
-                                        let status =
-                                            if let Ok(Async::Ready(t)) = c.runtime_state().poll() {
-                                                (*(t.status().clone()).to_string()).to_string()
-                                            } else {
-                                                "".to_string()
-                                            };
-                                        Module::new(c.name().to_string(), status)
-                                    })
-                                    .collect();
-                                f(mods) // changes depending on API call
-                            })
-                            .map_err(ErrorInternalServerError)
-                    })
-                    .into_future()
-                    .flatten(),
-            )
-        })
-        .unwrap_or_else(|err| {
-            Either::B(ok(HttpResponse::ServiceUnavailable()
-                .content_type("text/plain")
-                .body(format!("{:?}", err))))
-        });
-
-    Box::new(response)
-}
-
-fn module_response(mods: Vec<Module>) -> HttpResponse {
-    HttpResponse::Ok().body(format!("{:?}", mods))
-}
-
-fn health_response(mods: Vec<Module>) -> HttpResponse {
-    let mut device_status = health::Status::new();
-    let mut edge_agent = false;
-    let mut edge_hub = false;
-    let mut other = true;
-
-    for module in mods.iter() {
-        if module.name() == "edgeAgent" {
-            if module.status() == "running" {
-                edge_agent = true;
-            }
-        } else if module.name() == "edgeHub" {
-            if module.status() == "running" {
-                edge_hub = true;
-            }
-        } else {
-            if module.status() != "running" {
-                other = false;
-            }
-        }
-    }
-
-    device_status.set_iotedged();
-    device_status.set_edge_agent(edge_agent);
-    device_status.set_edge_hub(edge_hub);
-    device_status.set_other_modules(edge_agent && edge_hub && other);
-
-    let health = device_status.return_health();
-    HttpResponse::Ok().body(format!(
-        "Device health: {:?}\nDevice details: {:?}",
-        health, device_status
-    ))
-}
-
-fn resolve_and_tls_handshake(
-    to_socket_addrs: &impl std::net::ToSocketAddrs,
-    tls_hostname: &str,
-) -> Result<(), ActixError> {
-    let host_addr = to_socket_addrs
-        .to_socket_addrs()
-        .map_err(ErrorInternalServerError)?
-        .next()
-        .ok_or_else(|| "")
-        .map_err(ErrorInternalServerError)?;
-
-    let stream = TcpStream::connect_timeout(&host_addr, std::time::Duration::from_secs(10))
-        .map_err(ErrorInternalServerError)?;
-
-    let tls_connector = native_tls::TlsConnector::new().map_err(ErrorInternalServerError)?;
-
-    let _ = tls_connector
-        .connect(tls_hostname, stream)
-        .map_err(ErrorInternalServerError)?;
-
-    Ok(())
 }
 
 fn get_default_config_path() -> PathBuf {
